Refactoring. Working on minimizing memory usage.

+++ yuv/src/main/java/com/theeasiestway/yuv/entities/Frame.kt
@@ -7,13 +7,13 @@ import java.nio.ByteBuffer
 //
 
 abstract class Frame {
-    protected var nativePointer = -1
+    protected var nativePointer = -1L
     var width: Int = 0; protected set
     var height: Int = 0; protected set
 
     fun destroy() { destroy(nativePointer) }
     fun getBytes() = getBytes(nativePointer)
 
-    private external fun destroy(pointer: Int)
-    private external fun getBytes(pointer: Int): ByteBuffer
+    private external fun destroy(pointer: Long)
+    private external fun getBytes(pointer: Long): ByteBuffer
 }
